passport-telegram/pyproject.toml
[project]
name = "passport-telegram"
version = "0.1.0"
description = "Telegram bot adapter for passport-core"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "passport-core",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "python-telegram-bot[job-queue]>=21.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.6.4",
]

[project.scripts]
passport-telegram = "passport_telegram.cli:main"

[tool.uv.sources]
passport-core = { path = "../passport-core", editable = true }

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "A", "SIM", "RET", "PTH"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/passport_telegram"]
